+++ lib/active_interaction/errors.rb
@@ -2,9 +2,6 @@ module ActiveInteraction
   # Top-level error class. All other errors subclass this.
   Error = Class.new(StandardError)
 
-  # Raised when trying to run an empty pipeline.
-  EmptyPipelineError = Class.new(Error)
-
   # Raised if a class name is invalid.
   InvalidClassError = Class.new(Error)
 
@@ -28,6 +25,10 @@ module ActiveInteraction
 
   # Raised if there is no default value.
   NoDefaultError = Class.new(Error)
+
+  # @private
+  Interrupt = Class.new(::Interrupt)
+  private_constant :Interrupt
 
   # A small extension to provide symbolic error messages to make introspecting
   #   and testing easier.
@@ -79,6 +80,29 @@ module ActiveInteraction
     def clear
       symbolic.clear
       super
+    end
+
+    # Merge other errors into this one.
+    #
+    # @param other [Errors]
+    #
+    # @return [Errors]
+    def merge!(other)
+      other.symbolic.each do |attribute, symbols|
+        symbols.each do |symbol|
+          add_sym(attribute, symbol)
+        end
+      end
+
+      other.messages.each do |attribute, messages|
+        messages.each do |message|
+          unless added?(attribute, message)
+            add(attribute, message)
+          end
+        end
+      end
+
+      self
     end
   end
 end