+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # [Master][]
 
+- Add experimental composition implementation
+  (`ActiveInteraction::Base#compose`).
+- Remove `ActiveInteraction::Pipeline`.
+
 # [0.8.0][] (2013-11-14)
 
 - Add ability to document interactions and filters.